Update CSS for dark and light themes, Add HTML elements
Finished updating a rough draft of the CSS styles for both the light and
dark themes. Edited several elements' CSS for better accessibility (outline
when focused, adding an "active" class for future keyboard support via
Javascript).

Also added containers for a turn-tracker during a game and a winner modal
to display the winner's name as well as a button to play again.