Add win conditions and minor CSS tweaks
Added win conditions that are looked up after each turn, with a draw being
declared after 9 rounds if the gameboard state does not match any of the
win conditions.

Also made minor CSS tweaks to fix horizontal scrollbar issues and issues
when adjusting the screen size too low (<~350px so may not be an issue, but
wanted to fix it in case).