Update logic for gameboard array, Add cpu logic
Updated the logic for the gameBoard.board array. Previously the clicking
a cell on the screen would set that cell's textContent to that player's
marker and the board array would update based on that. Now clicking a cell
on the screen updates the board array (using the cell id to target the array
index), and the display is then updated based on the array.

Also began creating the logic for a cpu player to take their turn, which
will require testing upon completion.