Update winning conditions array, Tweak CSS, Update README
Updated the winning conditions array by turning it into a function that
returns the gameBoard.board indexes; previous attempts did not return the
updated indexes and instead returned empty (""). Adding the "return"
keyword fixed this issue.

Made small modifications to the CSS, adding a little color in the buttons in
an attempt to strive for passing  WCAG tests.

Finally, I updated the README in an attempt to create a more informative
ReadMe file, adding an image of the gameboard for decoration and adding a
"How to play" section.